Flower_Recognition_CNN
The "Flower Recognition CNN" is a project that uses a special kind of computer program to identify different types of flowers. In this project, you'll find information about why it was made and how it works. We used a collection of flower pictures to teach the program, and it can now recognize different flowers pretty well. If you want to try it or learn from it, we explain how to set it up and give some examples. We're open to others joining in and helping out, so feel free to get in touch with us if you're interested. This is your guide to understanding and using the project.